requirements.txt
certifi==2019.3.9
Django==2.2
django-crispy-forms==1.7.2
pytz==2018.9
sqlparse==0.3.0
